pyproject.toml
[tool.poetry]
name = "ai-clerk"
version = "0.1.0"
description = "AI Clerk - автоматизация OCR, NLP и заполнения форм"
authors = ["md <marceldiomare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
spacy = "^3.8.2"
transformers = "^4.46.3"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
